% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ROC.R
\name{add_ROC}
\alias{add_ROC}
\title{add_ROC}
\usage{
add_ROC(data, score_col)
}
\arguments{
\item{data}{A data frame containing a `TP` and score column.}

\item{score_col}{A character string naming the column to be used as a score column.}
}
\value{
A data frame containing the following columns: `True_Positive_Rate` and `False_Positive_Rate`.
}
\description{
Adds TPR and FPR columns to a data frame containing TP and a score metric (e.g. Bayes factors).
}
