% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_counts_QC.R
\name{assemble_counts_QC}
\alias{assemble_counts_QC}
\title{assemble_counts_QC}
\usage{
assemble_counts_QC(
  qc_metrics,
  counts,
  counts_norm,
  library,
  plasmid_sample_id,
  control_sample_id,
  treat_sample_id
)
}
\arguments{
\item{qc_metrics}{A data frame containing only sequencing-based QC metrics (rows as samples, QC metrics as columns).}

\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{counts_norm}{A data frame of normalized counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{library}{A data frame containing the library file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID.}

\item{plasmid_samp}{A character vector naming one or more plasmid samples.}

\item{control_samp}{A character vector naming one or more control samples.}

\item{treat_samp}{A character vector naming one or more treatment samples. If `NULL` then there are no treatment samples.}
}
\value{
A modified `qc_metrics` data frame containing all counts-based QC metrics (both experiment and sample-level).
}
\description{
Assembles counts-based QC metrics.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
