% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_logfc_QC.R
\name{assemble_logfc_QC}
\alias{assemble_logfc_QC}
\title{assemble_logfc_QC}
\usage{
assemble_logfc_QC(
  qc_metrics,
  screen_goal,
  lfc.ctrl_pl,
  lfc.treat_pl,
  lfc.ctrl_pl.repl,
  lfc.treat_pl.repl,
  library
)
}
\arguments{
\item{qc_metrics}{A data frame containing only sequencing-based QC metrics (rows as samples, QC metrics as columns).}

\item{screen_goal}{A character string naming the screen goal.}

\item{lfc.ctrl_pl}{A data frame of gRNA logFC data for Control vs Plasmid (gRNAs as rows).}

\item{lfc.treat_pl}{A data frame of gRNA logFC data for Treatment vs Plasmid (gRNAs as rows). Ignored, if `NULL`.}

\item{lfc.ctrl_pl.repl}{A data frame of gRNA logFC data for replicate-based Control vs Plasmid (gRNAs as rows).}

\item{lfc.treat_pl.repl}{A data frame of gRNA logFC data for replicate-based Treatment vs Plasmid (gRNAs as rows).}

\item{library}{A data frame containing the library file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID.}
}
\value{
A modified `qc_metrics` data frame containing all logFC-based QC metrics (both experiment and sample-level).
}
\description{
Assembles logFC-based QC metrics.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
