% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AUC.R
\name{calc_AUC}
\alias{calc_AUC}
\title{calc_AUC}
\usage{
calc_AUC(data, score_col, group = NULL)
}
\arguments{
\item{data}{A data frame containing a 'TP' column.}

\item{score_col}{A character string naming the column to be used as a score column.}

\item{group}{A character string naming a column by which to group 'data'. If `NULL` no grouping, defaults to `NULL`.}
}
\value{
A data frame.
}
\description{
Calculates AUROC in a data frame containing a 'TP' column.
}
