% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dispersion_adj_gRNA.R
\name{calc_dispersion_adj_gRNA}
\alias{calc_dispersion_adj_gRNA}
\title{calc_dispersion_adj_gRNA}
\usage{
calc_dispersion_adj_gRNA(counts, control_sample_id, treat_sample_id)
}
\arguments{
\item{counts}{A data frame of normalized counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{control_sample_id}{A character vector naming one or more control samples.}

\item{treat_sample_id}{A character vector naming one or more treatment samples. If `NULL` then there are no treatment samples.}
}
\description{
Calculates adjusted dispersion estimates at the gRNA level using a shrinkage estimator and returns the mean across all gRNAs for the Control vs Treatment comparison.
}
\references{
Wu, H. et al. 2013. A new shrinkage estimator for dispersion improves differential expression detection in RNA-seq data. Biostatistics 14: 232-243.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
