% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_euclid_dist_plasmid_PCA.R
\name{calc_euclid_dist_plasmid_PCA}
\alias{calc_euclid_dist_plasmid_PCA}
\title{calc_euclid_dist_plasmid_PCA}
\usage{
calc_euclid_dist_plasmid_PCA(
  counts,
  plasmid_sample_id,
  control_sample_id,
  treat_sample_id
)
}
\arguments{
\item{counts}{A data frame of normalized counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{plasmid_sample_id}{A character vector naming the plasmid sample(s).}

\item{control_sample_id}{A character vector naming the control sample(s).}

\item{treat_sample_id}{A character vector naming the treatment sample(s). Ignored if `NULL`.}
}
\value{
A data frame with the following column(s): `euclid_dist_plasmid_PCA.<comp>` where '<comp>' is one of 'control' or 'treat'.
}
\description{
Calculates a 2D-Euclidean distance from a set of given samples to the plasmid sample for the first 2 PCs of a PCA derived from the normalized counts. The minimum distance is returned for each comparison.
}
