% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_GC_percent_library.R
\name{calc_GC_percent_library}
\alias{calc_GC_percent_library}
\title{calc_GC_percent_library}
\usage{
calc_GC_percent_library(library)
}
\arguments{
\item{library}{A data frame containing the library file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID.}
}
\value{
A data frame with a column named `GC_percent`.
}
\description{
Calculates the GC percent of sgRNAs in a given library.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
