% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_GICC_gene_sets.R
\name{calc_GICC_gene_sets}
\alias{calc_GICC_gene_sets}
\title{calc_GICC_gene_sets}
\usage{
calc_GICC_gene_sets(
  counts,
  gene_sets,
  plasmid_sample_id,
  control_sample_id,
  treat_sample_id = NULL
)
}
\arguments{
\item{counts}{A data frame of normalized counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{gene_sets}{A named list of gene sets.}

\item{plasmid_sample_id}{A character vector naming one or more plasmid samples.}

\item{control_sample_id}{A character vector naming one or more control samples.}

\item{treat_sample_id}{A character vector naming one or more treatment samples. If `NULL` then there are no treatment samples. Defaults to `NULL`.}
}
\value{
A data frame with columns named `GICC.<comp>.<gene_set>` where '<comp>' is one of 'ctrl_plasmid', 'treat_plasmid', or 'treat_ctrl', and '<gene_set>' is the name of each gene set provided in the `gene_sets` list.
}
\description{
Calculates GICC for normalized counts given a list of gene sets.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
