% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gini_coefficient_counts.R
\name{calc_gini_coefficient_counts}
\alias{calc_gini_coefficient_counts}
\title{calc_gini_coefficient_counts}
\usage{
calc_gini_coefficient_counts(counts)
}
\arguments{
\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}
}
\value{
A data frame containing a `SampleName` column, and a column named `gini_coefficient_counts`.
}
\description{
Calculates the Gini coefficient of dispersion for sample gRNA count distributions.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
