% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inefficient_gRNA_count_ratios.R
\name{calc_inefficient_gRNA_count_ratios}
\alias{calc_inefficient_gRNA_count_ratios}
\title{calc_inefficient_gRNA_count_ratios}
\usage{
calc_inefficient_gRNA_count_ratios(counts, library)
}
\arguments{
\item{counts}{A data frame of normalized counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{library}{A data frame containing the library file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID.}
}
\value{
A data frame containing three columns: `SampleName`, `norm_counts_GCC_ratio` and `norm_counts_TT_ratio`.
}
\description{
Calculates the ratio of normalized counts for inefficient sgRNA PAM-proximal sequences relative to all guides (taking the median of each).
}
\references{
Graf, R. et al. (2019) sgRNA sequence motifs blocking efficient CRISPR/Cas9-mediated gene editing. Cell Rep 26: 1098-1103.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
