% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log2_fold_change_genes.R
\name{calc_log2_fold_change_genes}
\alias{calc_log2_fold_change_genes}
\title{calc_log2_fold_change_genes}
\usage{
calc_log2_fold_change_genes(logfc)
}
\arguments{
\item{logfc}{A data frame of log2 fold change data generated by `fgcQC::calc_log2_fold_change_gRNAs`.}
}
\value{
A data frame with a column named `log2FC_gene`.
}
\description{
Calculates median log2 fold changes for genes (taking as input output from `calc_log2_fold_change_gRNAs`).
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
