% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log2_fold_change_gRNAs.R
\name{calc_log2_fold_change_gRNAs}
\alias{calc_log2_fold_change_gRNAs}
\title{calc_log2_fold_change_gRNAs}
\usage{
calc_log2_fold_change_gRNAs(counts, ref, comp, pseudo_count = 5)
}
\arguments{
\item{counts}{A data frame of normalized sgRNA counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{ref}{A character vector naming the reference sample(s).}

\item{comp}{A character vector naming the comparison sample(s) to be contrasted against `ref` samples.}

\item{pseudo_count}{A small integer specifying a pseudo-count to add to the counts to avoid log2(0) cases. Defaults to 5.}
}
\value{
A data frame of log2 fold changes in which the columns are samples and the rows are sgRNAs.
}
\description{
Calculates log2 fold changes for sgRNAs from normalized counts. Median counts are taken across replicates.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
