% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_low_zero_count_plasmid_gRNAs.R
\name{calc_low_zero_count_plasmid_gRNAs}
\alias{calc_low_zero_count_plasmid_gRNAs}
\title{calc_low_zero_count_plasmid_gRNAs}
\usage{
calc_low_zero_count_plasmid_gRNAs(counts, plasmid_sample_id)
}
\arguments{
\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{plasmid_sample_id}{A character string naming the plasmid sample ID in the column names of `counts`. If several plasmid samples are given, the median count will be used.}
}
\value{
A data frame containing two colums `percent_low_count_plasmid_gRNAs` and `percent_zero_count_plasmid_gRNAs`.
}
\description{
Caclculate the percentage of low (<30) and zero count gRNAs in the plasmid sample - these should be rare.
}
