% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mahalanobis_dist_plasmid.R
\name{calc_mahalanobis_dist_plasmid}
\alias{calc_mahalanobis_dist_plasmid}
\title{calc_mahalanobis_dist_plasmid}
\usage{
calc_mahalanobis_dist_plasmid(
  counts,
  plasmid_sample_id,
  control_sample_id,
  treat_sample_id
)
}
\arguments{
\item{counts}{A data frame of normalized counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{plasmid_sample_id}{A character vector naming the plasmid sample(s).}

\item{control_sample_id}{A character vector naming the control sample(s).}

\item{treat_sample_id}{A character vector naming the treatment sample(s). Ignored if `NULL`.}
}
\value{
A data frame with the following column(s): `mahalanobis_dist_ratio.<comp>` where '<comp>' is one of 'ctrl_plasmid' or 'treat_plasmid'.
}
\description{
Calculates the Mahalanobis distance ratio for control vs plasmid and treatment vs plasmid.
}
