% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NNMD_robust.R
\name{calc_NNMD_robust}
\alias{calc_NNMD_robust}
\title{calc_NNMD_robust}
\usage{
calc_NNMD_robust(data, gene_col, lfc_col, essential_genes, nonessential_genes)
}
\arguments{
\item{data}{A data fram containing log fold change data (normalized in terms of library depth and any potential copy-number effects) for a set of genes.}

\item{gene_col}{A character string indicating which column contains gene names.}

\item{lfc_col}{A character string indicating which column contains $log_2$ fold change data for each gene.}

\item{essential_genes}{A vector of essential gene names.}

\item{nonessential_genes}{A vector of non-essential gene names.}
}
\value{
A single numeric value.
}
\description{
Calculates a robust equivalent of the Null-Normalized Mean Difference (NNMD; also known as Glass's delta) for the log fold change distributions of essential and non-essential genes - defined as the median difference divided by the median absolute deviation (MAD) of non-essential genes (the reference set).
}
