% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NNMD_robust_gene_sets.R
\name{calc_NNMD_robust_gene_sets}
\alias{calc_NNMD_robust_gene_sets}
\title{calc_NNMD_robust_gene_sets}
\usage{
calc_NNMD_robust_gene_sets(lfc_ctrl_pl, lfc_treat_pl, gene_sets)
}
\arguments{
\item{lfc_ctrl_pl}{A data frame of gRNA log fold change data for Control vs Plasmid.}

\item{lfc_treat_pl}{A data frame of gRNA log fold change data for Treatment vs Plasmid. If `NULL`, then ignored.}

\item{gene_sets}{A named list of gene sets.}
}
\value{
A data frame with columns named `NNMD_robust.<comp>.<gene_set>` where '<comp' is one of 'ctrl_plasmid' or 'treat_plasmid', and '<gene_set>' is the name of each gene set provided in the `gene_sets` list.
}
\description{
Calculates a Robust equivalent of the Null-Normalized Mean Difference (NNMD) for sets of genes (based on median difference divided by median absolute deviation of reference set).
}
