% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_percent_matching_gRNAs.R
\name{calc_percent_matching_gRNAs}
\alias{calc_percent_matching_gRNAs}
\title{calc_percent_matching_gRNAs}
\usage{
calc_percent_matching_gRNAs(b2f_samples, counts)
}
\arguments{
\item{b2f_samples}{A data frame of sample information derived from `bcl2fastq2` output - see `fgcQC::extract_b2f_json`. The `SampleName` column should have been replaced with sample names hailing from the analysis config.}

\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}
}
\value{
A data frame containing per-sample percent matching gRNA values in a column called `percent_reads_matching_gRNAs`.
}
\description{
Calculates the percentage of reads in each sample that can be assigned to any of the gRNA sequences.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
