% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_replicate_logfc_corr.R
\name{calc_replicate_logfc_corr}
\alias{calc_replicate_logfc_corr}
\title{calc_replicate_logfc_corr}
\usage{
calc_replicate_logfc_corr(logfc, suffix)
}
\arguments{
\item{logfc}{A data frame with columns `log2FC.repl_1` and `log2FC.repl_2`.}

\item{suffix}{A character string naming the comparison: one of `ctrl_plasmid` and `treat_plasmid`.}
}
\value{
A data frame with `repl_log2FC_pearson_corr.<suffix>` columns.
}
\description{
Calculates logFC replicate Pearson correlations in accordance with recommendations by Hanna & Doench (2020).
}
\references{
Hanna, R. E. and Doench, J. G. 2020. Design and analysis of CRISPR-Cas experiments. Nature Biotechnology.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
