% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanr.lfc.fgc_0008.R
\docType{data}
\name{cleanr.lfc.fgc_0008}
\alias{cleanr.lfc.fgc_0008}
\title{cleanr.lfc.fgc_0008}
\description{
CRISPRcleanR corrected log fold change data for an HT-29 QC screen performed at the FGC.
}
\keyword{data}
