% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_qc_metrics.R
\docType{data}
\name{config_qc_metrics}
\alias{config_qc_metrics}
\title{config_qc_metrics}
\description{
A data frame providing the column names of the 'qc' section of the analysis config.
}
\keyword{data}
