% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_gene_sets.R
\docType{data}
\name{crispr_gene_sets}
\alias{crispr_gene_sets}
\title{crispr_gene_sets}
\description{
Sets of essential, non-essential, and functional classes of genes for use in CRISPR pooled-screen QC.
}
\keyword{data}
