% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_analysis_comparisons.R
\name{extract_analysis_comparisons}
\alias{extract_analysis_comparisons}
\title{extract_analysis_comparisons}
\usage{
extract_analysis_comparisons(json_list)
}
\arguments{
\item{json_list}{A list returned by `fgcQC::read_analysis_config_json`.}
}
\value{
A data frame containing sample comparisons.
}
\description{
Extracts analysis comparisons from output generated by `fgcQC::read_analysis_config_json`.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
