% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_b2f_json.R
\name{extract_b2f_json}
\alias{extract_b2f_json}
\title{extract_b2f_json}
\usage{
extract_b2f_json(path)
}
\arguments{
\item{path}{A valid path to a 'Stats.json' file produced by `bcl2fastq2`.}
}
\value{
A list containing the following elements:
`flowcell` - a data frame containing flowcell-level data.
`run_info` - a data frame containing read lengths.
`summary` - a data frame containing run-level QC metrics.
`samples` - a data frame containing sample-level QC metrics.
}
\description{
Extracts Sequencing QC metrics from a `bcl2fastq2` 'Stats.json' file.
}
\examples{
\dontrun{jj <- extract_b2f_json('Stats.json')}
}
