% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_proximal_4bases_PAM.R
\name{extract_proximal_4bases_PAM}
\alias{extract_proximal_4bases_PAM}
\title{extract_proximal_4bases_PAM}
\usage{
extract_proximal_4bases_PAM(library)
}
\arguments{
\item{library}{A data frame containing the library file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID.}
}
\value{
A data frame with a column named `proximal_4bases_pam`.
}
\description{
Extracts the proximal 4 bases upstream of the PAM in a sgRNA library data frame into "GCC", "TT", or "Other" (Graf et al. 2019).
}
\references{
Graf, R. et al. (2019) sgRNA sequence motifs blocking efficient CRISPR/Cas9-mediated gene editing. Cell Rep 26: 1098-1103.
}
