% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcQC-package.R
\docType{package}
\name{fgcQC-package}
\alias{fgcQC}
\alias{fgcQC-package}
\title{fgcQC: QC metric calculations for CRISPR screen data produced at the FGC}
\description{
Implementations of QC metric calcluations pertaining to sequencing, laboratory, and analysis metrics; both standard and novel metrics are included.
}
\details{
Provides QC metric calculation functions for FGC CRISPR pooled screen data. The main entry point is `QC_fgc_crispr_data` and this function takes an analysis config JSON file, count data, bagel output, sequencing metrics (bcl2fastq), and a gRNA library file. QC data is produced for a single comparison in the analysis config.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
\keyword{internal}
