% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PRROC.R
\name{get_PRROC}
\alias{get_PRROC}
\title{get_PRROC}
\usage{
get_PRROC(data, score_col, group_col = NULL)
}
\arguments{
\item{data}{A data frame containing a 'TP' logical column indicating which genes are true positives.}

\item{score_col}{A character string naming a column containing a score metric.}

\item{group_col}{A character string naming the grouping column of 'data'. If `NULL` then no group, defaults to `NULL`.}
}
\value{
A data frame containing the following columns: `Precision`, `Recall`, `Sensitivity_FDR_10pct`, `Sensitivity_FDR_10pct`, and `AUPrRc`.
}
\description{
Returns precision and recall values for a data frame containing TP and a score metric (e.g. Bayes factor), or the AUPrRc.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
