% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bcl2fastq_config_qc.R
\name{merge_bcl2fastq_config_qc}
\alias{merge_bcl2fastq_config_qc}
\title{merge_bcl2fastq_config_qc}
\usage{
merge_bcl2fastq_config_qc(b2f, qc_config, comparisons)
}
\arguments{
\item{b2f}{A data frame of `bcl2fastq` sequencing metrics.}

\item{qc_config}{A data frame of QC metrics taken from the 'qc' section of the analysis config JSON.}

\item{comparisons}{A data frame of comparisons taken from the 'comparisons' section of the analysis config JSON.}
}
\value{
A data frame.
}
\description{
Merge bcl2fastq sequencing metrics with the QC section in the analysis config.
}
