% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_bcl2fastq_json.R
\name{mock_bcl2fastq_json}
\alias{mock_bcl2fastq_json}
\title{mock_bcl2fastq_json}
\usage{
mock_bcl2fastq_json(sample_ids)
}
\arguments{
\item{sample_ids}{A character vector naming the sample IDs to be mocked.}
}
\value{
A file path to a temporary mock 'bcl2fastq' JSON file.
}
\description{
Mock bcl2fastq data for a specified number of samples with given sample IDs (to match an analysis config), write to a temporary file, and return the temp file path.
}
