% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_bcl2fastq_template.R
\docType{data}
\name{mock_bcl2fastq_template}
\alias{mock_bcl2fastq_template}
\title{mock_bcl2fastq_template}
\description{
A template for mock bcl2fastq data - to be manipulated and written to a temp file.
}
\keyword{data}
