% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_library_cleanr_tsv.R
\name{mock_library_cleanr_tsv}
\alias{mock_library_cleanr_tsv}
\title{mock_library_cleanr_tsv}
\usage{
mock_library_cleanr_tsv(combined_counts)
}
\arguments{
\item{combined_counts}{A path to a valid combined counts csv file (produced by the AZ-CRUK CRISPR analysis pipeline).}
}
\value{
A file path to a temporary mock 'cleanr' file.
}
\description{
Mock a 'cleanr.tsv' gRNA library annotation file, write to a temporary file, and return the temporary file path.
}
