% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_missing_FGC_data.R
\name{mock_missing_FGC_data}
\alias{mock_missing_FGC_data}
\title{mock_missing_FGC_data}
\usage{
mock_missing_FGC_data(analysis_config, combined_counts, bcl2fastq, library)
}
\arguments{
\item{analysis_config}{A path to a valid analysis config JSON file to be used as input for the AZ-CRUK CRISPR analysis pipeline.}

\item{combined_counts}{A path to a valid combined counts csv file (produced by the AZ-CRUK CRISPR analysis pipeline).}

\item{bcl2fastq}{A character string giving one or more paths to valid `bcl2fastq2` summary output JSON files (paths separated by commas). `NULL` if missing.}

\item{library}{A valid path to a library tsv file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID (produced by the AZ-CRUK CRISPR reference data generation pipeline). `NULL` if missing.}
}
\value{
A named list of temporary file names pointing to mocked data.
}
\description{
Function to write out mock data that is missing from the inputs to `QC_fgc_crispr_data`. The mock data is written to temporary files that are deleted when the `R` session ends.
}
