% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fgcQC.R
\name{print.fgcQC}
\alias{print.fgcQC}
\title{print.fgcQC}
\usage{
\method{print}{fgcQC}(x, ...)
}
\arguments{
\item{x}{An object of class `fgcQC`.}

\item{...}{Other named arguments to be passed to `print`.}
}
\value{
Prints a summary of the object to the console.
}
\description{
S3 method for generic function `print` for printing summaries of objects of class `fgcQC` to the console.
}
