% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inefficient_gRNA_logfc.R
\name{calc_inefficient_gRNA_logfc}
\alias{calc_inefficient_gRNA_logfc}
\title{calc_inefficient_gRNA_logfc}
\usage{
calc_inefficient_gRNA_logfc(logfc, library)
}
\arguments{
\item{logfc}{A data frame of log2 fold change data for each sample in the study (samples as columns, gRNAs as rows) as generated by `fgcQC::calc_log2_fold_change_gRNAs`.}

\item{library}{A data frame containing the library file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID.}
}
\value{
A data frame containing three columns: `SampleName`, `log2FC_GCC_ratio` and `log2FC_TT_ratio`.
}
\description{
Calculates the ratio of normalized counts for inefficient sgRNA PAM-proximal sequences relative to all guides (taking the median of each).
}
\references{
Graf, R. et al. (2019) sgRNA sequence motifs blocking efficient CRISPR/Cas9-mediated gene editing. Cell Rep 26: 1098-1103.
}
