% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GICC.R
\name{GICC}
\alias{GICC}
\title{GICC}
\usage{
GICC(data, return.cov = TRUE, all.gicc = FALSE, conf.level = 0.95)
}
\arguments{
\item{data}{A matrix or data frame of multivariate data in which the first column contains between-subject identifiers and the remaining columns are the variables measured on these subjects.}

\item{return.cov}{A logical indicating whether to return the covariance matrix or not. Defaults to \code{TRUE}.}

\item{all.gicc}{A logical indicating whether to return all of Ahrens GICC measures. Defaults to \code{FALSE}.}

\item{conf.level}{A numeric value between 0 and 1 indicating the desired confidence level for the GICC confidence interval.}
}
\value{
A list with the following components:
\code{GICC} - A list with the various GICC measures from Ahrens original paper. Use \code{GICC.1} for the one advocated by Ahrens.
\code{num.var} - An integer giving the number of variables.
\code{num.samps} - An integer giving the number of samples.
\code{num.subj} - An integer giving the number of subjects.
\code{num.reps} - An integer giving the number of replicates.
\code{conf.level} - The confidence level for the CI.
\code{balanced.design} - A logical indicating if the design is balanced or not.
}
\description{
Calculates Ahrens (1976) multivariate generalized intraclass correlation coefficient (GICC).
}
\references{
Ahrens, H. (1976). Multivariate variance-covariance components (MVCC) and generalized intraclass correlation coefficient (GICC). Biom. Z. 18, 527-533.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
