% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_fgc_crispr_data.R
\name{QC_fgc_crispr_data}
\alias{QC_fgc_crispr_data}
\title{QC_fgc_crispr_data}
\usage{
QC_fgc_crispr_data(
  analysis_config,
  combined_counts,
  bagel_ctrl_plasmid,
  bcl2fastq,
  library,
  comparison_name,
  output,
  norm_method = "median_ratio",
  html_report = FALSE,
  historical_data = NULL
)
}
\arguments{
\item{analysis_config}{A path to a valid analysis config JSON file to be used as input for the AZ-CRUK CRISPR analysis pipeline.}

\item{combined_counts}{A path to a valid combined counts csv file (produced by the AZ-CRUK CRISPR analysis pipeline).}

\item{bagel_ctrl_plasmid}{A path to a valid Bagel tsv results file for Control vs Plasmid.}

\item{bcl2fastq}{A character string giving one or more paths to valid `bcl2fastq2` summary output JSON files (paths separated by commas).}

\item{library}{A valid path to a library tsv file in which the first column gives the sgRNA sequence and the second column gives the sgRNA ID (produced by the AZ-CRUK CRISPR reference data generation pipeline).}

\item{comparison_name}{A character string naming a single comparison to extract QC data for (should correspond to the comparison name used in the analysis config JSON file).}

\item{output}{A character string giving an output file name for the csv results. If `NULL`, do not write out any results.}

\item{norm_method}{A character string naming a normalization method for the count data. Can be `median_ratio` or `relative`. Defaults to `median_ratio`.}

\item{html_report}{A logical indicating whether an HTML report for the focal comparison should be produced. Defaults to `FALSE`.}

\item{historical_data}{If `html_report` is `TRUE` then a data frame of historical QC data to be used in the report. Set to `NULL` if `html_report` is `FALSE`. Defaults to `NULL`.}
}
\value{
A list containing the following elements:
`qc_metrics` - A data frame containing QC metrics as columns and samples as rows; this data will also be written to the `output` file, if not `NULL`.
`comparisons` - A data frame of samples belonging to the focal comparison.
`seq_metrics` - A data frame of CI sequencing metrics at both flowcell and sample levels.
`log2FC` - A list containing logFC data frames at the gRNA and gene level.
}
\description{
Produce Quality Control (QC) metrics from the following FGC CRISPR screen data inputs: an analysis config JSON file, a combined counts csv file, a Bagel results tsv file (Control vs Plasmid), a gRNA library tsv file ('cleanr.tsv'), and one or more bcl2fastq2 output JSON files ('Stats.json'). QC output will be returned for a single named comparison.
}
